Rudimentary Cli tool to update CRLs
