+++ examples/Udap.Pki.Cli/CrlType.cs
@@ -1,0 +1,7 @@
+﻿namespace Udap.Pki.Cli;
+
+public enum CrlType
+{
+    CA,
+    SubCA
+}
