+++ examples/Udap.Pki.Cli/PublicCertStoreConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace Udap.Pki.Cli;
+
+public class PublicCertStoreConfig
+{
+    public string CrlBucket { get; set; } = string.Empty;
+    public string CrlPath { get; set; } = string.Empty;
+}
